images working

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
@@ -50,14 +51,35 @@ class PostController extends Controller
             
             'title' => 'required|max:255',
             'body' => 'required|max:255',
+            'cover_image' => 'image|nullable|max:1999'
             
         ]);
 
-        $a = new Post;
-        $a->user_id = auth()->user()->id;
-        $a->title = $validatedData['title'];
-        $a->body = $validatedData['body'];
-        $a->save();
+         //file upload
+         if($request->hasFile('cover_image')){
+            // get filename with whatever extension it has
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just the filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //get only the ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store in db, makes filename unique incase identical images uploaded
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //upload
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
+        
+
+
+        //make new post
+        $post = new Post;
+        $post->user_id = auth()->user()->id;
+        $post->title = $validatedData['title'];
+        $post->body = $validatedData['body'];
+        $post->cover_image = $fileNameToStore;
+        $post->save();
 
         session()->flash('message', 'Post was created.');
         return redirect()->route('posts.index');
@@ -83,7 +105,14 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+
         $post = Post::findOrFail($id);
+        //check for user
+        if(auth()->user()->id !== $post->user_id){
+            session()->flash('error', 'Unauthorized');
+            return redirect('/posts');
+           
+        }
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -103,10 +132,28 @@ class PostController extends Controller
             
         ]);
 
+ //file upload
+ if($request->hasFile('cover_image')){
+    // get filename with whatever extension it has
+    $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+    //get just the filename
+    $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+    //get only the ext
+    $extension = $request->file('cover_image')->getClientOriginalExtension();
+    //filename to store in db, makes filename unique incase identical images uploaded
+    $fileNameToStore = $filename.'_'.time().'.'.$extension;
+    //upload
+    $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+}
+
+
         $post = Post::findOrFail($id);
-        $post->user_id = auth()->user()->id;
+       
         $post->title = $validatedData['title'];
         $post->body = $validatedData['body'];
+        if($request->hasFile('cover_image')){
+        $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         session()->flash('message', 'Post was edited.');
@@ -122,6 +169,19 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
+        //check for user
+        if(auth()->user()->id !== $post->user_id){
+            session()->flash('error', 'Unauthorized');
+            return redirect('/posts');
+           
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            //delete 
+            Storage::delete('public/cover_images/'.$post->cover_image);
+
+        }
+
         $post->delete();
 
         return redirect()->route('posts.index')->with('message', 'Post was deleted.');
